@@ -1,6 +1,6 @@
 ﻿import { InternetService } from "./InternetService";
 import { Injectable } from "@angular/core";
-import { Subject } from "rxjs/Subject";
+import { Subject } from "rxjs";
 
 @Injectable()
 export class AuthService {
