@@ -1,28 +1,28 @@
-﻿import { ToastsManager } from 'ng2-toastr/ng2-toastr';
+﻿
 import {
     ViewContainerRef, Injectable
 } from '@angular/core';
 
 @Injectable()
 export class ToastService {
-    constructor(public toastr: ToastsManager) { }
+    //constructor(public toastr: ToastsManager) { }
     success(msg:string) {
-        this.toastr.success(msg);
+        //this.toastr.success(msg);
     }
 
     info(msg: string) {
-        this.toastr.info(msg);
+        //this.toastr.info(msg);
     }
 
     warning(msg: string) {
-        this.toastr.warning(msg);
+        //this.toastr.warning(msg);
     }
 
     error(msg: string) {
-        this.toastr.error(msg);
+        //this.toastr.error(msg);
     }
 
     setVCR(vcr: ViewContainerRef) {
-        this.toastr.setRootViewContainerRef(vcr);
+        //this.toastr.setRootViewContainerRef(vcr);
     }
 }
