@@ -3,7 +3,7 @@ import { CropperSettings } from "ng2-img-cropper";
 import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material";
 
 @Component({
-    selector: 'AvatarPicker',
+    selector:'AvatarPicker',
     template: ` 
               <h3 mat-dialog-title>Pick an Image</h3>
                <mat-dialog-content>
@@ -11,7 +11,7 @@ import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material";
                     <img *ngIf="data.image" [src]="data.image" [width]="cropperSettings.croppedWidth" [height]="cropperSettings.croppedHeight">
                 </mat-dialog-content>
                 <mat-dialog-actions>
-                    <button type="button" mat-dialog-close class="btn btn-secondary" data-dismiss="modal"  (click)="onNoClick()>Close</button>
+                    <button type="button" mat-dialog-close class="btn btn-secondary" data-dismiss="modal"  (click)="onNoClick()">Close</button>
                     <button type="button" class="btn btn-primary" (click)="onDoneClick()">Done</button>
                 </mat-dialog-actions>`,
 })
@@ -21,7 +21,6 @@ export class AvatarPicker {
 
     constructor(public dialogRef: MatDialogRef<AvatarPicker>,
         @Inject(MAT_DIALOG_DATA) public dataSent: any) {
-
         this.cropperSettings = new CropperSettings();
         this.cropperSettings.width = 100;
         this.cropperSettings.height = 100;
@@ -34,10 +33,12 @@ export class AvatarPicker {
     }
 
     onNoClick(): void {
+        console.log("done with popup");
         this.dialogRef.close();
     }
 
     onDoneClick(): void {
+        console.log("Cancelling with popup");
         this.dialogRef.close(this.data.image);
     }
 
