@@ -1,14 +1,14 @@
 ﻿export class videoData {
-    videoId: Number;
-    folderName : string;
-    fileName : string;
-    poster: string;
-    format: string;
-    title: string;
-    description: string;
-    userId: string;
-    userFirstName: string;
-    userDistrict: string;
-    numberOfLikes: Number;
-    numberOfComments: Number;
+    videoId: Number=0;
+    folderName : string = "";
+    fileName: string = "";
+    poster: string = "";
+    format: string = "";
+    title: string = "";
+    description: string = "";
+    userId: string = "";
+    userFirstName: string = "";
+    userDistrict: string = "";
+    numberOfLikes: Number = 0;
+    numberOfComments: Number = 0;
 }
