@@ -7,9 +7,9 @@
     <base href="~/" />
 
     <link rel="stylesheet" href="~/dist/vendor.css" asp-append-version="true" />
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/angularjs-toaster/2.2.0/toaster.min.css" rel="stylesheet" />
-    @*<link href="node_modules/@angular/material/prebuilt-themes/indigo-pink.css" rel="stylesheet" />*@
+    @*<link href="https://cdnjs.cloudflare.com/ajax/libs/angularjs-toaster/2.2.0/toaster.min.css" rel="stylesheet" />*@
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <link href="~/indigo-pink.css" rel="stylesheet" />
 </head>
 <body>
     @RenderBody()
